taxi-back/requirements.txt
Django
psycopg2-binary
pycryptodome
djangorestframework
requests
django-cors-headers
